+++ action-platformer/main.js
@@ -35,12 +35,12 @@ scene("main", () => {
   ];
 
   const map = addLevel([
-    "                                            ",
-    "                                            ",
-    "                                            ",
-    "                     w                      ",
-    "              =================   w         ",
-    "==============----__----_-_---_========  ===",
+    "                                                                                                                                       ",
+    "                                                                                                                                       ",
+    "                                                            w      s                                                                   ",
+    "                     w                                  ================       w                                   s                   ",
+    "              =================   w          w   =======-___---≈≈-≈--_≈≈===========     w       s       w     ==============  ====     ",
+    "==============----__----_-_---_========  ========--_-__-________---_____--__--___-_===========================--≈_-_≈-_-≈--_  -_≈-=====",
   ], {
     width: 32,
     height: 32,
@@ -62,6 +62,10 @@ scene("main", () => {
       }),
       layer("enviroment"),
     ],
+    "≈": () => [
+      sprite("platform3"),
+      layer("enviroment"),
+    ],
     "w": () => [
       sprite("fire-worm", {
         anim: "walking",
@@ -78,6 +82,25 @@ scene("main", () => {
         reachLimit: false,
         dir: -1,
         speed: 15,
+        death: false,
+      }
+    ],
+    "s": () => [
+      sprite("spider", {
+        anim: "walk",
+        flipX: false,
+      }),
+      area({width: 25, height: 20, offset: {x: 0, y: 5}}),
+      origin("center"),
+      body(),
+      layer("enemies"),
+      "enemy",
+      "spider",
+      {
+        limit: {max: 0, min: 0},
+        reachLimit: false,
+        dir: -1,
+        speed: 60,
         death: false,
       }
     ],
@@ -167,7 +190,7 @@ scene("main", () => {
           arrow.move(-250, 0);
         }
       })
-      arrow.collides("enemy", (e) => {
+      arrow.collides("worm", (e) => {
         play("arrow-hit", {
           volume: 0.2,
         });
@@ -176,6 +199,16 @@ scene("main", () => {
         play("worm-death", {volume: 0.2});
         e.play("death")
         wait(1, () => destroy(e));
+      })
+      arrow.collides("spider", (e) => {
+        play("arrow-hit", {
+          volume: 0.2,
+        });
+        destroy(arrow);
+        e.death = true;
+        play("spider-death", {volume: 0.2});
+        e.play("death");
+        wait(0.6, () => {destroy(e);})
       })
     });
   }
@@ -229,15 +262,17 @@ scene("main", () => {
     }
   };
 
-  player.collides("enemy", () => {
-    player.hit = true;
-    play("player-hit", {volume: 0.2})
-    if(player.direction > 0){
-      player.moveBy(-50, 0);
-    }else {
-      player.moveBy(50, 0);
+  player.collides("enemy", (e) => {
+    if(!e.death){
+      player.hit = true;
+      play("player-hit", {volume: 0.2})
+      if(player.direction > 0){
+        player.moveBy(-50, 0);
+      }else {
+        player.moveBy(50, 0);
+      }
+      player.hurt(1);
     }
-    player.hurt(1);
   })
 
   player.on("death", () => {
@@ -255,31 +290,61 @@ scene("main", () => {
       player.falling = true;
     }
 
-    if(player.pos.x > 180){
+    if(player.pos.x > 4200){
+      go("end-demo");
+    }
+
+    if(player.pos.x > 180 && player.pos.x < 4139){
       camPos({x: player.pos.x, y: 90})
     }
+
+    console.log(player.pos.x)
   })
 
   const worms = get("worm");
-  for(const worm of worms){
-    worm.limit = {max: worm.pos.x + 50, min: worm.pos.x - 50};
-    worm.action(() => {
-      if(!worm.death){
-        if(Math.floor(worm.pos.x) > worm.limit.max){
-          worm.reachLimit = true;
-          worm.dir = -1;
-          worm.flipX(true);
+  for(const w of worms){
+    w.limit = {max: w.pos.x + 50, min: w.pos.x - 50};
+    w.action(() => {
+      if(!w.death){
+        if(Math.floor(w.pos.x) > w.limit.max){
+          w.reachLimit = true;
+          w.dir = -1;
+          w.flipX(true);
         }
-        if(Math.floor(worm.pos.x) < worm.limit.min){
-          worm.reachLimit = false;
-          worm.dir = 1;
-          worm.flipX(false);
+        if(Math.floor(w.pos.x) < w.limit.min){
+          w.reachLimit = false;
+          w.dir = 1;
+          w.flipX(false);
         }
     
-        if(worm.reachLimit){
-          worm.move(worm.speed * worm.dir, 0);
+        if(w.reachLimit){
+          w.move(w.speed * w.dir, 0);
         }else {
-          worm.move(worm.speed * worm.dir, 0)
+          w.move(w.speed * w.dir, 0)
+        }
+      }
+    })
+  }
+  const spiders = get("spider");
+  for(const s of spiders){
+    s.limit = {max: s.pos.x + 70, min: s.pos.x - 70};
+    s.action(() => {
+      if(!s.death){
+        if(Math.floor(s.pos.x) > s.limit.max){
+          s.reachLimit = true;
+          s.dir = -1;
+          s.flipX(false);
+        }
+        if(Math.floor(s.pos.x) < s.limit.min){
+          s.reachLimit = false;
+          s.dir = 1;
+          s.flipX(true);
+        }
+    
+        if(s.reachLimit){
+          s.move(s.speed * s.dir, 0);
+        }else {
+          s.move(s.speed * s.dir, 0)
         }
       }
     })
@@ -296,6 +361,14 @@ scene("gameover", () => {
   keyPress("enter", () => {
     go("main");
   })
+})
+
+scene("end-demo", () => {
+  add([
+    text("END OF THE DEMO, STAY TUNED TO SEE THE FINAL PROJECT", {size: 15}),
+    pos(width()/2, height()/2),
+    origin("center"),
+  ])
 })
 
 go("main");